fix: defines are placed with correct context
Fixes #26

The problem of swapping the define keywords was that I wasn't checking
the lexical scope of the insertion point - which results in things like
reported error - an number wasn't be cast as an int.

Now when a define is found, it's swapped in, the statement rebuilt, then
the bytes generated, turned into text, *then* re-tokenised. This ensures
the scopes and tokens are all absolutely correct.